GpioControlWidget_Qt5/requirements.txt
numpy
pandas
PyQt5
PyQt5-Qt5
PyQt5_sip
rpi-lgpio

